src/main/java/eatitproject/eatitproject/services/CommandeServiceImpl.java
package eatitproject.eatitproject.services;


import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import eatitproject.eatitproject.models.Commande;
import eatitproject.eatitproject.repository.CommandeRepository;



@Service
public class CommandeServiceImpl implements CommandeService {

    @Autowired
    private CommandeRepository commandeRepository;
    
    @Override
    public List<Commande> getAllCommandes() {
        return commandeRepository.findAll();
    }

    @Override
    public Commande getCommandeById(Long id) {
        return commandeRepository.findById(id).orElse(null);
    }

    @Override
    public Commande updateCommande(Commande commande) {
        return commandeRepository.save(commande);
    }

    @Override
    public void deleteCommande(Long id) {
        commandeRepository.deleteById(id);
    }

	@Override
	public Commande createCommande(Commande commande) {
		return commandeRepository.save(commande);
	}

}
